my_app123/coffee_users_postgres/alembic/versions/4c2f484afc35_first.py
"""first

Revision ID: 4c2f484afc35
Revises: 
Create Date: 2025-05-26 11:44:27.539419

"""
from typing import Sequence, Union

from alembic import op
import sqlalchemy as sa


# revision identifiers, used by Alembic.
revision: str = '4c2f484afc35'
down_revision: Union[str, None] = None
branch_labels: Union[str, Sequence[str], None] = None
depends_on: Union[str, Sequence[str], None] = None


def upgrade() -> None:
    """Upgrade schema."""
    pass


def downgrade() -> None:
    """Downgrade schema."""
    pass
