my_app123/coffee_users_postgres/requirements.txt
Flask
Flask-SQLAlchemy
requests
psycopg2
